% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf4_baixar_cjsg_id.R
\name{trf4_baixar_cjsg_id}
\alias{trf4_baixar_cjsg_id}
\title{Baixar julgados de segundo grau conforme id}
\usage{
trf4_baixar_cjsg_id(id = NULL, diretorio = ".")
}
\arguments{
\item{id}{Id do processo. Pode ser o id completo ou o número.}

\item{diretorio}{Diretório}
}
\value{
htmls
}
\description{
Baixar julgados de segundo grau conforme id
}
\details{
Há uma sequência de ids que podem ser baixados. Vc pode formá-la, se quiser.
}
