% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf4_baixar_cposg_id.R
\name{trf4_baixar_cposg_id}
\alias{trf4_baixar_cposg_id}
\title{Baixar cposg conforme id}
\usage{
trf4_baixar_cposg_id(id = NULL, diretorio = ".")
}
\arguments{
\item{id}{Id do processo. Pode ser completo ou seomente os números.}

\item{diretorio}{Diretório}
}
\value{
htmls
}
\description{
Baixar cposg conforme id
}
\details{
Há uma sequência de ids que podem ser baixados. Vc pode formá-la, se quiser.
}
