% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf4_baixar_inteiro_teor.R
\name{trf4_baixar_inteiro_teor}
\alias{trf4_baixar_inteiro_teor}
\title{Baixa inteiro teor de acórdãos ou decisões monocráticas}
\usage{
trf4_baixar_inteiro_teor(id = NULL, url = NULL, diretorio = NULL)
}
\arguments{
\item{id}{Id do processo. Pode ser completo ou seomente os números.}

\item{url}{Url do processo}

\item{diretorio}{Diretório onde armazenar os arquivos}
}
\value{
html ou pdf
}
\description{
Baixa inteiro teor de acórdãos ou decisões monocráticas
}
