% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf4_cjsg.R
\name{trf4_cjsg}
\alias{trf4_cjsg}
\title{Obtêm os ids dos processos a partir da busca livre}
\usage{
trf4_cjsg(
  livre = "",
  data_inicial = "",
  data_final = "",
  tipo_data = c("DEC", "PUB")
)
}
\arguments{
\item{livre}{Busca livre.}

\item{data_inicial}{Data inicial no formato dd/mm/yyyy}

\item{data_final}{Data final no formato dd/mm/yyyy}

\item{tipo_data}{Buscar por data de decisão "DEC" ou data de
publicação "PUB".}
}
\value{
Vetor de ids
}
\description{
Obtêm os ids dos processos a partir da busca livre
}
\details{
Após obter os ids, use a função trf4_baixar_cjsg para
baixar por ids. Limite a busca por data inicial e final
}
