% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf4_ler_cjsg.R
\name{trf4_ler_cjsg}
\alias{trf4_ler_cjsg}
\title{Ler julgados de segundo grau do TRF4}
\usage{
trf4_ler_cjsg(arquivos = NULL, diretorio = ".", wide = TRUE)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Diretório, se não informar arquivos}

\item{wide}{Converte para formato largo}
}
\value{
tibble
}
\description{
Ler julgados de segundo grau do TRF4
}
