% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf4_ler_dados_cposg2.R
\name{trf4_ler_dados_cposg2}
\alias{trf4_ler_dados_cposg2}
\title{Ler dados da consulta processual do 2º grau do TRF4}
\usage{
trf4_ler_dados_cposg2(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Diretório onde se encontram os arquivos}
}
\value{
tibble
}
\description{
Ler dados da consulta processual do 2º grau do TRF4
}
