% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf4_ler_docs_html.R
\name{trf4_ler_docs_html}
\alias{trf4_ler_docs_html}
\title{Lê julgados}
\usage{
trf4_ler_docs_html(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar se arquivos não forem informados}
}
\value{
tibble
}
\description{
Lê julgados
}
