% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf4_ler_partes_docs_html.R
\name{trf4_ler_partes_docs_html}
\alias{trf4_ler_partes_docs_html}
\title{Lê partes dos htmls dos julgados}
\usage{
trf4_ler_partes_docs_html(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar se não informar arquivos}
}
\value{
tibble
}
\description{
Lê partes dos htmls dos julgados
}
