% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf4_ler_url_inteiro_teor.R
\name{trf4_ler_url_inteiro_teor}
\alias{trf4_ler_url_inteiro_teor}
\title{Lê urls dos inteiros teores do TRF4}
\usage{
trf4_ler_url_inteiro_teor(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Se não informar vetor de arquivos}
}
\value{
tibble
}
\description{
Lê urls dos inteiros teores do TRF4
}
