% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf4_ler_url_processo.R
\name{trf4_ler_url_processo}
\alias{trf4_ler_url_processo}
\title{Lê urls dos  processos de segundo grau do TRF4 a partir dos hmtls do cjsg}
\usage{
trf4_ler_url_processo(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Se não informar vetor de arquivos}
}
\value{
tibble
}
\description{
Lê urls dos  processos de segundo grau do TRF4 a partir dos hmtls do cjsg
}
